ref(preprod): Move SnapshotsToolbar to its own file
Eliminates ~90 lines of transitive mocks from the snapshot tests by
importing from the standalone module instead of snapshotMainContent.
